refactor: Manual 관련 ui 분리
